generated_code/fixtures/tests/test_generatefakedata_fixture.py
from code_autoeval.llm_model.utils.preprocess_code_before_exec import PreProcessCodeBeforeExec
from code_autoeval.llm_model.utils.prompting.system_prompts import SystemPrompts
from code_autoeval.llm_model.utils.model_response.stream_response import StreamResponse
from code_autoeval.llm_model.utils.preprocess_code_before_exec import PreProcessCodeBeforeExec
from code_autoeval.llm_model.utils.prompting.system_prompts import SystemPrompts
from code_autoeval.llm_model.utils.model_response.stream_response import StreamResponse
import pytest
from generated_code.fixtures.fixtures.generatefakedata_fixture import fixture_mock_generatefakedata
from code_autoeval.llm_model.utils.generate_fake_data import GenerateFakeData
from pathlib import Path
from typing import Any, Dict, Optional
def test_mock_generatefakedata(fixture_mock_generatefakedata):
    assert isinstance(fixture_mock_generatefakedata, GenerateFakeData)
    assert hasattr(fixture_mock_generatefakedata, 'async_generate_fake_data')
    assert callable(fixture_mock_generatefakedata.async_generate_fake_data)
    assert hasattr(fixture_mock_generatefakedata, 'generate_fake_data')
    assert callable(fixture_mock_generatefakedata.generate_fake_data)
    assert hasattr(fixture_mock_generatefakedata, 'coverage_result')
    assert hasattr(fixture_mock_generatefakedata, 'file_path')
    assert isinstance(fixture_mock_generatefakedata.file_path, Path)
    assert hasattr(fixture_mock_generatefakedata, 'test_file_path')
    assert isinstance(fixture_mock_generatefakedata.test_file_path, Path)
    assert hasattr(fixture_mock_generatefakedata, 'absolute_path_from_root')
    assert isinstance(fixture_mock_generatefakedata.absolute_path_from_root, Path)
    assert hasattr(fixture_mock_generatefakedata, 'unique_imports_dict')
    assert isinstance(fixture_mock_generatefakedata.unique_imports_dict, Dict)
    assert isinstance(fixture_mock_generatefakedata, StreamResponse)
    assert isinstance(fixture_mock_generatefakedata, PreProcessCodeBeforeExec)
    assert isinstance(fixture_mock_generatefakedata, SystemPrompts)